backend/requirements.txt
flask
scikit-learn
joblib
pandas
numpy
pymongo
python-dotenv
gunicorn